+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## **WORK IN PROGRESS**
+* (Gerhard Steinwedel) **NEW**: Added temperature-controlled automatic plan duration adjustment
+* (Gerhard Steinwedel) **FIXED**: Reset automation duration states when stopping a run
+
 ## 0.2.28 (2026-07-23)
 * (Gerhard Steinwedel) **NEW**: Added Hydrawise valve control and auto-discovery through ioBroker.hydrawise
 * (Gerhard Steinwedel) **ENHANCED**: Documented all ioBroker adapter data points and their usage in the README